:recycle: :sparkles: :bug: :bulb: 降低代码异味，修复若干 bug
