+++ app.vue
@@ -21,6 +21,14 @@
 		return langs[locale.value as keyof typeof langs] ?? locale.value;
 	});
 
+	/** 是否启用彩色侧边栏 */
+	const coloredSideBar = computed(() => {
+		return (
+			appSettings.coloredSideBar
+			// && Theme.actualPalette.value // FIXME 不知道为什么 Theme.actualPalette.value 在 SSR 的时候会报错？
+		);
+	});
+
 	useHead({
 		htmlAttrs: {
 			lang: langTag,
@@ -31,15 +39,7 @@
 			{ "http-equiv": "X-UA-Compatible", content: "IE=Edge,chrome=1" },
 			{ name: "viewport", content: "width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1" },
 			{ name: "renderer", content: "webkit" },
-			{
-				name: "theme-color",
-				content: computed(() => {
-					return (
-						appSettings.coloredSideBar
-						// && Theme.actualPalette.value // FIXME 不知道为什么 Theme.actualPalette.value 在 SSR 的时候会报错？
-					);
-				}),
-			},
+			{ name: "theme-color", content: coloredSideBar },
 			{ name: "description", content: manifest.description },
 			{ name: "keywords", content: "视频,弹幕,字幕,音频,歌词,相簿,相册,照片,视频网站,弹幕视频,二次元,动漫,动画,音乐,动漫音乐,音MAD,AMV,MAD,ANIME,ACG,NOVA" },
 			// 以下内容为各种苹果私有属性。
