+++ components/Login/LoginWindow.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-	import { CUSTOM_THEME_COLOR, DEFAULT_THEME_COLOR, SYSTEM_THEME, coloredSidebarCookieKey, customThemeColorCookieKey, themeColorCookieKey, themeTypeCookieKey, cookieBinding, isOfflineSettingsCookieKey, sharpAppearanceModeCookieKey, flatAppearanceModeCookieKey } from "~/modules/theme/cookieBinding";
-	import { PALETTE_LIST } from "~/modules/theme/types";
+	import { cookieBinding } from "modules/theme/cookieBinding";
+	import { saveUserSetting2BrowserCookie } from "modules/theme/composables";
 
 	const props = defineProps<{
 		/** 已打开，单向绑定使用。 */
@@ -58,28 +58,12 @@
 					open.value = false;
 					selfUserInfoStore.isLogined = true;
 
+					// 登陆后，将用户设置存储到 cookie，然后调用 cookieBinding 从 cookie 中获取样式设置并追加到 dom 根节点
 					const userSettings = await api.user.getUserSettings();
-					
-					// 登陆后，将用户设置存储到 cookie
-					const currentThemeType = userSettings?.userSettings?.themeType || SYSTEM_THEME;
-					const themeColor = userSettings?.userSettings?.themeColor ? (PALETTE_LIST as unknown as string[]).includes(userSettings.userSettings.themeColor) ? userSettings.userSettings.themeColor : CUSTOM_THEME_COLOR : DEFAULT_THEME_COLOR;
-					const customerThemeColor = userSettings?.userSettings?.themeColor || "";
-					const isColoredSidebar = userSettings?.userSettings?.coloredSideBar || false;
-					const isSharpAppearanceMode = userSettings?.userSettings?.sharpAppearanceMode || false;
-					const isFlatAppearanceMode = userSettings?.userSettings?.flatAppearanceMode || false;
-					// HACK 7 在此处添加
-
-					const userSettingsCookieBasicOption = `; expires=${new Date("9999/9/9").toUTCString()}; path=/; SameSite=Strict`;
-					document.cookie = `${isOfflineSettingsCookieKey}=false${userSettingsCookieBasicOption}`;
-					if (currentThemeType) document.cookie = `${themeTypeCookieKey}=${currentThemeType}${userSettingsCookieBasicOption}`;
-					if (themeColor) document.cookie = `${themeColorCookieKey}=${themeColor}${userSettingsCookieBasicOption}`;
-					if (customerThemeColor) document.cookie = `${customThemeColorCookieKey}=${customerThemeColor}${userSettingsCookieBasicOption}`;
-					if (isColoredSidebar !== undefined && isColoredSidebar !== null) document.cookie = `${coloredSidebarCookieKey}=${isColoredSidebar}${userSettingsCookieBasicOption}`;
-					if (isSharpAppearanceMode !== undefined && isSharpAppearanceMode !== null) document.cookie = `${sharpAppearanceModeCookieKey}=${isSharpAppearanceMode}${userSettingsCookieBasicOption}`;
-					if (isFlatAppearanceMode !== undefined && isFlatAppearanceMode !== null) document.cookie = `${flatAppearanceModeCookieKey}=${isFlatAppearanceMode}${userSettingsCookieBasicOption}`;
-					// HACK 8 在此处添加
-
+					saveUserSetting2BrowserCookie(userSettings);
 					cookieBinding();
+
+					// 触发用户登录事件
 					useEvent("user:login", true);
 				} else
 					useToast(t.toast.login_failed, "error");
