+++ composables/api/User/UserControllerDto.d.ts
@@ -12,7 +12,7 @@ export type UserRegistrationRequestDto = {
 	// // TODO 用户创建时间
 	// /** 用户创建时间 */
 	// userCreateDate: number;
-}
+};
 
 /**
  * 用户注册的返回参数
@@ -26,7 +26,7 @@ export type UserRegistrationResponseDto = {
 	token?: string;
 	/** 附加的文本消息 */
 	message?: string;
-}
+};
 
 /**
  * 用户登录提交的参数
@@ -36,7 +36,7 @@ export type UserLoginRequestDto = {
 	email: string;
 	/** 在前端已经 Hash 过一次的的密码 */
 	passwordHash: string;
-}
+};
 
 /**
  * 用户登录的返回参数
@@ -54,7 +54,7 @@ export type UserLoginResponseDto = {
 	passwordHint?: string;
 	/** 附加的文本消息 */
 	message?: string;
-}
+};
 
 /**
  * 验证用户邮箱是否存在提交的参数
@@ -62,7 +62,7 @@ export type UserLoginResponseDto = {
 export type UserExistsCheckRequestDto = {
 	/** 用户邮箱 */
 	email: string;
-}
+};
 
 /**
  * 验证用户邮箱是否已经存在的返回参数
@@ -74,7 +74,7 @@ export type UserExistsCheckResponseDto = {
 	exists: boolean; // WARN: 用户已存在或查询失败时都会返回 true
 	/** 附加的文本消息 */
 	message?: string;
-}
+};
 
 /**
  * 用户更改邮箱的请求的参数
@@ -88,7 +88,7 @@ export type UpdateUserEmailRequestDto = {
 	newEmail: string;
 	/** 经过一次 Hash 的用户密码 */
 	passwordHash: string;
-}
+};
 
 /**
  * 用户更改邮箱返回的参数
@@ -98,7 +98,7 @@ export type UpdateUserEmailResponseDto = {
 	success: boolean;
 	/** 附加的文本消息 */
 	message?: string;
-}
+};
 
 /**
  * 等待被 Hash 的密码和用户信息
@@ -108,7 +108,7 @@ export type BeforeHashPasswordDataType = {
 	email: string;
 	/** 在前端已经 Hash 过一次的的密码 */
 	passwordHash: string;
-}
+};
 
 
 
@@ -120,7 +120,7 @@ export type UserLabel = {
 	id: number;
 	/** 标签名 */
 	labelName: string;
-}
+};
 
 /**
  * 用户的关联账户
@@ -130,7 +130,7 @@ export type UserLinkAccounts = {
 	accountType: string;
 	/** 关联账户唯一标识 */
 	accountUniqueId: string;
-}
+};
 
 /**
  * 用户的关联网站
@@ -140,7 +140,7 @@ export type UserWebsite = {
 	websiteName: string;
 	/** 关联网站 URL */
 	websiteUrl: string;
-}
+};
 
 /**
  * 更新或创建用户信息时的请求参数
@@ -166,7 +166,7 @@ export type UpdateOrCreateUserInfoRequestDto = {
 	userLinkAccounts?: UserLinkAccounts[];
 	/** 用户的关联网站 */
 	userWebsite?: UserWebsite;
-}
+};
 
 
 /**
@@ -179,7 +179,7 @@ export type UpdateOrCreateUserInfoResponseDto = {
 	message?: string;
 	/** 请求结果 */
 	result?: {} & UpdateOrCreateUserInfoRequestDto;
-}
+};
 
 
 
@@ -191,7 +191,7 @@ export type GetSelfUserInfoRequestDto = {
 	uid: number;
 	/** 用户的身分令牌 */
 	token: string;
-}
+};
 
 /**
  * 获取当前登录的用户信息的请求响应
@@ -203,7 +203,7 @@ export type GetSelfUserInfoResponseDto = {
 	message?: string;
 	/** 请求结果 */
 	result?: { uid?: number } & UpdateOrCreateUserInfoRequestDto;
-}
+};
 
 /**
  * 通过 UID 获取用户信息的请求载荷
@@ -211,7 +211,7 @@ export type GetSelfUserInfoResponseDto = {
 export type GetUserInfoByUidRequestDto = {
 	/** 目标用户的 UID */
 	uid: number;
-}
+};
 
 /**
  * 通过 UID 获取用户信息的请求响应
@@ -236,7 +236,7 @@ export type GetUserInfoByUidResponseDto = {
 		/** 用户的个人标签 */
 		label?: UserLabel[];
 	};
-}
+};
 
 /**
  * 通过 UID 和 TOKEN 校验用户的返回结果
@@ -248,7 +248,7 @@ export type CheckUserTokenResponseDto = {
 	message?: string;
 	/** 用户校验结果，用户正常为 true，非法用户为 false */
 	userTokenOk?: boolean;
-}
+};
 
 /**
  * 用户登出的响应
@@ -258,7 +258,7 @@ export type UserLogoutResponseDto = {
 	success: boolean;
 	/** 附加的文本消息 */
 	message?: string;
-}
+};
 
 /**
  * 获取用于用户上传头像的预签名 URL, 上传限时 60 秒
@@ -270,7 +270,7 @@ export type GetUserAvatarUploadSignedUrlResultDto = {
 	userAvatarUploadSignedUrl?: string;
 	/** 附加的文本消息 */
 	message?: string;
-}
+};
 
 /**
  * 用户关联账户的隐私设置
@@ -280,7 +280,7 @@ type UserLinkAccountsPrivacySettingDto = {
 	accountType: string;
 	/** 显示方式 - 非空 - 允许的值有：{public: 公开, following: 仅关注, private: 隐藏} */
 	privacyType: 'public' | 'following' | 'private';
-}
+};
 
 /**
  * 基础用户个性设置类型
@@ -326,12 +326,12 @@ export type BasicUserSettingsDto = {
 	userWebsitePrivacySetting?: 'public' | 'following' | 'private';
 	/** 用户关联账户的隐私设置 */
 	userLinkAccountsPrivacySetting?: UserLinkAccountsPrivacySettingDto[];
-}
+};
 
 /**
  * 获取用于渲染页面的用户设定的请求参数
  */
-export type GetUserSettingsRequestDto = {} & GetSelfUserInfoRequestDto
+export type GetUserSettingsRequestDto = {} & GetSelfUserInfoRequestDto;
 
 /**
  * 获取用于渲染页面的用户设定的请求响应
@@ -343,13 +343,13 @@ export type GetUserSettingsResponseDto = {
 	userSettings?: { uid: number; editDateTime: number } & BasicUserSettingsDto;
 	/** 附加的文本消息 */
 	message?: string;
-}
+};
 
 
 /**
  * 更新或创建用户设定的请求参数
  */
-export type UpdateOrCreateUserSettingsRequestDto = {} & BasicUserSettingsDto
+export type UpdateOrCreateUserSettingsRequestDto = {} & BasicUserSettingsDto;
 
 /**
  * 更新或创建用户设定的请求响应
@@ -361,4 +361,4 @@ export type UpdateOrCreateUserSettingsResponseDto = {
 	userSettings?: { uid: number; editDateTime: number } & BasicUserSettingsDto;
 	/** 附加的文本消息 */
 	message?: string;
-}
+};
